NJIT CS635 Programming project
The program does the following:
Read a 3d model described in a Wavefront .obj file and scale the x, y
and z axes by user inputted factors. Write an output .obj file with
the object scaled.
Look up online the format of a Wavefront .obj file to see how to
extract the vertex coordinates. When you write the output file, write
it without the surface normal vectors (which will be invalid after
the axis scaling).